Create README.md
Single Vendor a white label food delivering solution that allow owners to easily manage their deliveries. It can help to intuitively and instantly automate your deliveries ,and handle the logistics. Our food delivery solution provides the capability of order management, as well as separate application for the rider and customer.